Moves histogram count to int32
Towards https://github.com/giantswarm/giantswarm/issues/6130

We use a 64 bit int to store the number of observations in a bucket,
which is excessive for number of occurrences in a bucket.
Changing to int32 helps reduce memory usage, as we use a number
of histograms (e.g: w/ histogramvec) in some projects.